Friendly bot V2 (update 1.0)
Just making it easier to see the answer to your jokes, I'm still trying to find a way to make it so that whatever joke you get, you get that exact answer, but for now (and probably forever) there will not be the exact answer to your joke. Please recommend any things that I should add to it in the comments.